break url handler into package, remove third party scripts
