+++ reader.py
@@ -1,15 +1,44 @@
-#!/home/multimedia/Desktop/reader/venv/bin/python
+#!/usr/bin/env python
 
 import ndef
 import nfc
-import urihandler
+import urihandlers
+
+from urllib.parse import urlparse
+from urihandlers import AlbumHandler
+
+
+handlers = {
+    AlbumHandler.Scheme: AlbumHandler("/home/multimedia/SeaDrive/Shared with groups/Media/Music")
+}
+
+
+def handle_uris(uris):
+    previous = None
+    parsed_uris = []
+    for uri in uris:
+        parsed = urlparse(uri)
+        if previous is not None and previous.scheme != parsed.scheme:
+            raise Exception(f"multiple schemes may not be used (found {parsed.scheme} and {previous.scheme}")
+        if parsed.scheme not in handlers:
+            raise Exception("provided scheme unsupported")
+        previous = parsed
+        parsed_uris.append(parsed)
+    handlers[parsed.scheme].handle_uris(parsed_uris)
+
+
+def validate_uri(uri):
+    parsed = urlparse(uri)
+    if parsed.scheme not in handlers:
+        raise Exception(f"scheme {parsed.scheme} not in handlers: {repr(handlers.keys())}")
+    handlers[parsed.scheme].validate_uri(parsed)
 
 
 def get_uris(tag):
     if not tag.ndef:
         raise Exception("received empty ndef")
     if len(tag.ndef.records) == 0:
-        raise Exception(f"received tag with 0 records")
+        raise Exception("received tag with 0 records")
     uris = []
     for index, record in enumerate(tag.ndef.records):
         if not isinstance(record, ndef.uri.UriRecord):
@@ -18,6 +47,7 @@ def get_uris(tag):
         urihandler.validate_uri(uri)
         uris.append(uri)
     return uris
+
 
 def get_on_connect(device):
     def on_connect(tag):
